benchmarker/scripts/run_weaviate_query_agent.py
import weaviate
from weaviate.agents.query import QueryAgent

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
WEAVIATE_URL = os.getenv("WEAVIATE_URL")
WEAVIATE_API_KEY = os.getenv("WEAVIATE_API_KEY")

weaviate_client = weaviate.connect_to_weaviate_cloud(
    cluster_url=WEAVIATE_URL,
    auth_credentials=weaviate.auth.AuthApiKey(WEAVIATE_API_KEY),
    headers={"X-OpenAI-Api-Key": OPENAI_API_KEY}
)

qa = QueryAgent(
    client=weaviate_client,
    collections=["Menus"]
)

response = qa.run("How many vegetarian dishes are there?")
print(response)