benchmarker/pyproject.toml
[project]
name = "benchmarker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.24",
    "weaviate-client>=4.10.4",
]

[tool.uv.sources]
weaviate-client = { git = "https://github.com/cdpierse/weaviate-python-client.git", rev = "main" }
